refactor: replace D2Shim, UserSettingsProvider with CurrentUserProvider
CurrentUserProvider returns both user and user settings data.
D2Shim is not needed anymore as we don't need to pass d2 to FileMenu, we
pass currentUser from the provider instead, which is used in
OpenFileMenu (d2.currentUser was used before).
The local UserSettingsProvider became redundant since the settings info
is returned by CurrentUserProvider in analytics.