fix: reset app when New is clicked
Before it only happened on the 2nd click on New.
I'm not sure about the logic of the navigation, while debugging I
noticed that previousLocation is always undefined on the next
navigation, so the whole thing seems to work only because we use
location state which covers the cases where the URL remains the same,
but we still want to refresh the page.
The other case is to load the visualization when the location changes,
and that check is always true due to previousLocation being undefined.

This fixes the issue, and everything seems to work, but the logic needs
to be reviewed.